requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
geographiclib==2.1
geopy==2.4.1
idna==3.11
kafka-python==2.3.0
logging==0.4.9.6
psutil==7.1.3
psycopg2-binary==2.9.11
py4j==0.10.9.7
pyspark==3.4.0
requests==2.32.5
urllib3==2.6.2
pandas==2.3.3
hdfs==2.7.3
